+++ src/ResourceManager/DataLakeStore/Commands.DataLakeStore/ChangeLog.md
@@ -18,6 +18,8 @@
         - Additional information about change #1
 -->
 ## Current Release
+* Fix debugging when DebugPreference is set from powershell command line
+* Update example for Set-AzureRmDataLakeStoreItemAcl
 
 ## Version 6.1.1
 * Updated all help files to include full parameter types and correct input/output types.