+++ src/ResourceManager/Resources/Commands.Resources/ChangeLog.md
@@ -18,6 +18,7 @@
         - Additional information about change #1
 -->
 ## Current Release
+* Fix example in New-AzureRmResourceGroupDeployment
 
 ## Version 6.3.0
 * Updated help files to include full parameter types and correct input/output types.